push

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -14,7 +14,7 @@ int main(void)
 
 	long d = 0;
 
-	int k;
+	long k;
 
 	for (; d <= 4000000;)
 	{
